+++ cpplib/math/polyNTT.hpp
@@ -6,7 +6,7 @@ template<int N>
 class PolyBaseNTT : public PolyBase<MInt<N>> {
  protected:
   PolyBaseNTT mul(const PolyBaseNTT& rhs) const {
-    int n = this->size(), m = rhs.size(), tot = std::max(1, n + m - 1);
+    int n = (int)this->size(), m = rhs.size(), tot = std::max(1, n + m - 1);
     int sz = 1 << std::__lg(tot * 2 - 1);
     std::vector<MInt<N>> A = *this, B = rhs;
     A.resize(sz); B.resize(sz);
